made HM and log 4wd common
