+++ gbpservice/nfp/configurator/drivers/firewall/vyos/vyos_fw_driver.py
@@ -18,7 +18,6 @@ from oslo_log import log as logging
 from oslo_serialization import jsonutils
 
 from gbpservice.nfp.configurator.drivers.base import base_driver
-from gbpservice.nfp.configurator.lib import constants as common_const
 from gbpservice.nfp.configurator.lib import fw_constants as const
 
 LOG = logging.getLogger(__name__)
@@ -30,79 +29,14 @@ configuration requests.
 """
 
 
-class FwGenericConfigDriver(object):
+class FwGenericConfigDriver(base_driver.BaseDriver):
     """
     Driver class for implementing firewall configuration
     requests from Orchestrator.
     """
 
-    def __init__(self, conf):
-        self.conf = conf
-        self.timeout = const.REST_TIMEOUT
-
-    def _configure_log_forwarding(self, url, mgmt_ip, port):
-        """ Configures log forwarding IP address in Service VMs.
-
-            :param url: url format that is used to invoke the Service VM API
-            :param mgmt_ip: management IP of the Service VM
-            :param port: port that is listened to by the Service VM agent
-
-            Returns: SUCCESS/Error msg
-
-        """
-
-        url = url % (mgmt_ip, port, 'configure-rsyslog-as-client')
-
-        log_forward_ip_address = self.conf.log_forward_ip_address
-        if not log_forward_ip_address:
-            msg = ("Log forwarding IP address not configured "
-                   "for service at %s." % mgmt_ip)
-            LOG.info(msg)
-            return common_const.UNHANDLED
-
-        data = dict(
-                server_ip=log_forward_ip_address,
-                server_port=self.conf.log_forward_port,
-                log_level=self.conf.log_level)
-        data = jsonutils.dumps(data)
-
-        msg = ("Initiating POST request to configure log forwarding "
-               "for service at: %r" % mgmt_ip)
-        LOG.info(msg)
-        try:
-            resp = requests.post(url, data, timeout=self.timeout)
-        except requests.exceptions.ConnectionError as err:
-            msg = ("Failed to establish connection to service at: "
-                   "%r for configuring log forwarding. ERROR: %r" %
-                   (mgmt_ip, str(err).capitalize()))
-            LOG.error(msg)
-            return msg
-        except requests.exceptions.RequestException as err:
-            msg = ("Unexpected ERROR happened while configuring "
-                   "log forwarding for service at: %r. "
-                   "ERROR: %r" %
-                   (mgmt_ip, str(err).capitalize()))
-            LOG.error(msg)
-            return msg
-
-        try:
-            result = resp.json()
-        except ValueError as err:
-            msg = ("Unable to parse response of configure log forward API, "
-                   "invalid JSON. URL: %r. %r" % (url, str(err).capitalize()))
-            LOG.error(msg)
-            return msg
-        if not result['status']:
-            msg = ("Error configuring log forwarding for service "
-                   "at %s. URL: %r. Reason: %s." %
-                   (mgmt_ip, url, result['reason']))
-            LOG.error(msg)
-            return msg
-
-        msg = ("Successfully configured log forwarding for "
-               "service at %s." % mgmt_ip)
-        LOG.info(msg)
-        return const.STATUS_SUCCESS
+    def __init__(self):
+        pass
 
     def _configure_static_ips(self, resource_data):
         """ Configure static IPs for provider and stitching interfaces
@@ -131,7 +65,7 @@ class FwGenericConfigDriver(object):
         mgmt_ip = resource_data['mgmt_ip']
 
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'add_static_ip')
         data = jsonutils.dumps(static_ips_info)
 
@@ -190,14 +124,14 @@ class FwGenericConfigDriver(object):
 
         try:
             result_log_forward = self._configure_log_forwarding(
-                const.request_url, mgmt_ip, const.CONFIGURATION_SERVER_PORT)
+                const.request_url, mgmt_ip, self.port)
         except Exception as err:
             msg = ("Failed to configure log forwarding for service at %s. "
                    "Error: %s" % (mgmt_ip, err))
             LOG.error(msg)
             return msg
         else:
-            if result_log_forward == common_const.UNHANDLED:
+            if result_log_forward == const.UNHANDLED:
                 pass
             elif result_log_forward != const.STATUS_SUCCESS:
                 msg = ("Failed to configure log forwarding for service at %s. "
@@ -227,7 +161,7 @@ class FwGenericConfigDriver(object):
             stitching_mac=resource_data['stitching_mac'])
 
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT, 'add_rule')
+                                   self.port, 'add_rule')
         data = jsonutils.dumps(rule_info)
         msg = ("Initiating POST request to add persistent rule to primary "
                "service at: %r" % mgmt_ip)
@@ -286,7 +220,7 @@ class FwGenericConfigDriver(object):
         mgmt_ip = resource_data['mgmt_ip']
 
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'del_static_ip')
         data = jsonutils.dumps(static_ips_info)
 
@@ -363,7 +297,7 @@ class FwGenericConfigDriver(object):
         msg = ("Initiating DELETE persistent rule.")
         LOG.info(msg)
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'delete_rule')
 
         try:
@@ -417,7 +351,7 @@ class FwGenericConfigDriver(object):
         # REVISIT(VK): This was all along bad way, don't know why at all it
         # was done like this.
 
-        url = const.request_url % (mgmt_ip, const.CONFIGURATION_SERVER_PORT,
+        url = const.request_url % (mgmt_ip, self.port,
                                    'add-source-route')
         active_configured = False
         route_info = []
@@ -484,7 +418,7 @@ class FwGenericConfigDriver(object):
         # REVISIT(VK): This was all along bad way, don't know why at all it
         # was done like this.
         active_configured = False
-        url = const.request_url % (mgmt_ip, const.CONFIGURATION_SERVER_PORT,
+        url = const.request_url % (mgmt_ip, self.port,
                                    'delete-source-route')
         route_info = []
         for source_cidr in source_cidrs:
@@ -529,7 +463,7 @@ initialized. Also, only this driver class is exposed to the agent.
 """
 
 
-class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
+class FwaasDriver(FwGenericConfigDriver):
     service_type = const.SERVICE_TYPE
     service_vendor = const.VYOS
 
@@ -537,8 +471,9 @@ class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
         self.conf = conf
         self.timeout = const.REST_TIMEOUT
         self.host = self.conf.host
+        self.port = const.CONFIGURATION_SERVER_PORT
         self.context = context.get_admin_context_without_session()
-        super(FwaasDriver, self).__init__(conf)
+        super(FwaasDriver, self).__init__()
 
     def _get_firewall_attribute(self, firewall):
         """ Retrieves management IP from the firewall resource received
@@ -622,7 +557,7 @@ class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
         LOG.debug(msg)
         mgmt_ip = self._get_firewall_attribute(firewall)
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'configure-firewall-rule')
         msg = ("Initiating POST request for FIREWALL ID: %r Tenant ID:"
                " %r. URL: %s" % (firewall['id'], firewall['tenant_id'], url))
@@ -680,7 +615,7 @@ class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
 
         mgmt_ip = self._get_firewall_attribute(firewall)
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'update-firewall-rule')
         msg = ("Initiating UPDATE request. URL: %s" % url)
         LOG.info(msg)
@@ -714,7 +649,7 @@ class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
 
         mgmt_ip = self._get_firewall_attribute(firewall)
         url = const.request_url % (mgmt_ip,
-                                   const.CONFIGURATION_SERVER_PORT,
+                                   self.port,
                                    'delete-firewall-rule')
         msg = ("Initiating DELETE request. URL: %s" % url)
         LOG.info(msg)
@@ -763,19 +698,3 @@ class FwaasDriver(FwGenericConfigDriver, base_driver.BaseDriver):
             self._print_exception('Failure', resp.status_code, url,
                                   'create', resp.content)
             return const.STATUS_ERROR
-
-    def configure_healthmonitor(self, context, kwargs):
-        """Overriding BaseDriver's configure_healthmonitor().
-           It does netcat to CONFIGURATION_SERVER_PORT  8888.
-           Configuration agent runs inside service vm.Once agent is up and
-           reachable, service vm is assumed to be active.
-           :param context - context
-           :param kwargs - kwargs coming from orchestrator
-
-           Returns: SUCCESS/FAILED
-
-        """
-        ip = kwargs.get('mgmt_ip')
-        port = str(const.CONFIGURATION_SERVER_PORT)
-        command = 'nc ' + ip + ' ' + port + ' -z'
-        return self._check_vm_health(command)
