+++ gbpservice/nfp/configurator/lib/fw_constants.py
@@ -27,6 +27,7 @@ STATUS_DELETED = "DELETED"
 STATUS_UPDATED = "UPDATED"
 STATUS_ERROR = "ERROR"
 STATUS_SUCCESS = "SUCCESS"
+UNHANDLED = "UNHANDLED"
 
 request_url = "http://%s:%s/%s"
 SUCCESS_CODES = [200, 201, 202, 203, 204]
